[BUGFIX] Add table name
It is strongly recommended to define a table_name for all namespaced models. Not defining a table_name may cause generated table names to be too long and may not be supported by your current database engine. The generated naming scheme will also change when upgrading to SilverStripe 5.0 and potentially break.